backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
requests==2.32.4
pycryptodome==3.23.0
PyYAML==6.0.2
websockets==15.0.1
openpyxl
xlrd>=2.0.1
pandas
SQLAlchemy>=2.0
duckdb>=1.0.0
joblib
pydantic==2.10.4
pydantic-settings==2.7.0
python-multipart
APScheduler==3.10.4
simhash==2.1.2
beautifulsoup4==4.12.2
feedparser==6.0.10
tenacity==9.0.0
